pyproject.toml
[tool.poetry]
name = "any_tx_builder"
version = "0.1.0"
description = "Any-Blockchain transaction builder"
authors = ["Exo <0xexomonk@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
web3 = "^7.2.0"
eth-account = "^0.13.3"
eth-abi = "^5.1.0"
eth-utils = "^5.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
